Add matchbox_simple_demo with connect_native code + wasm support
