+++ matchbox_simple_demo/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "matchbox_simple_demo"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+futures-timer = { version = "3", features = ["wasm-bindgen"] }
+log = { version = "0.4", default-features = false }
+matchbox_socket = { path = "../matchbox_socket" }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+console_error_panic_hook = "0.1.7"
+console_log = "0.2.0"
+futures = { version = "0.3", default-features = false }
+wasm-bindgen-futures = "0.4.29"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+futures = "0.3"
+tokio = "1.12"
+pretty_env_logger = "0.4"
