+++ matchbox_socket/Cargo.toml
@@ -46,7 +46,3 @@ async-tungstenite = { version = "0.15", default-features = false, features = [ "
 webrtc = { version = "0.4", default-features = false }
 bytes = { version = "1.1", default-features = false }
 async-compat = { version = "0.2.1", default-features = false }
-
-[dev-dependencies]
-tokio = "1.12"
-pretty_env_logger = "0.4"
